the Base class is moved to another class
